+++ package.json
@@ -34,7 +34,7 @@
   "homepage": "https://github.com/znck/rollup-plugin-vue#readme",
   "dependencies": {
     "@babel/runtime": "^7.0.0-beta.39",
-    "@vue/component-compiler": "^3.1.0",
+    "@vue/component-compiler": "3.1.1",
     "@vue/component-compiler-utils": "^1.2.1",
     "debug": "^2.6.0",
     "hash-sum": "^1.0.2",
